+++ tests/client/e2eTests/helpers/TaxonomyHelper.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var util = require('util');
-var shortid = require('shortid');
+var nanoid = require('nanoid').nanoid;
 var e2e = require('@openveo/test').e2e;
 var Helper = e2e.helpers.Helper;
 
@@ -35,11 +35,11 @@ util.inherits(TaxonomyHelper, Helper);
  */
 TaxonomyHelper.prototype.getAddExample = function() {
   return {
-    id: shortid.generate(),
+    id: nanoid(),
     name: 'Taxonomy example',
     tree: [
       {
-        id: shortid.generate(),
+        id: nanoid(),
         items: [],
         title: 'Term example'
       }