+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - **mustache** has been upgraded from 3.2.1 to **4.2.0**
 - **nopt** has been upgraded from 4.0.3 to **5.0.0**
 - **semver** has been upgraded from 7.1.3 to **7.3.5**
+- **shortid** has been replaced by nanoid
 
 # 9.0.2 / 2020-11-05